add react router and default navigation with sample components
