+++ frontend/package.json
@@ -4,17 +4,18 @@
   "proxy": "http://django:8000",
   "private": true,
   "dependencies": {
-    "axios": "0.21.2",
-    "@testing-library/react": "11.2.7",
     "@testing-library/jest-dom": "5.12.0",
-    "@testing-library/user-event": "10.3.5",
+    "@testing-library/react": "11.2.7",
     "@testing-library/user-event": "13.1.9",
+    "axios": "0.21.2",
     "dotenv": "10.0.0",
     "eslint-config-react-app": "6.0.0",
     "eslint-plugin-flowtype": "5.7.2",
     "react": "17.0.2",
     "react-dom": "17.0.2",
-    "react-scripts": "3.4.4",
+    "react-router": "^6.3.0",
+    "react-router-dom": "^6.3.0",
+    "react-scripts": "^5.0.1",
     "typescript": "4.2.4"
   },
   "scripts": {
