+++ frontend/src/App.js
@@ -1,38 +1,46 @@
 import React from 'react';
 import logo from './logo.svg';
 import axios from 'axios';
+import { Routes, Route, BrowserRouter as Router, Link } from 'react-router-dom';
 import './App.css';
-
-require('dotenv').config()
-
-function handleSubmit(event) {
-  const text = document.querySelector('#char-input').value
-
-  axios
-    .get(`/char_count?text=${text}`).then(({data}) => {
-      document.querySelector('#char-count').textContent = `${data.count} characters!`
-    })
-    .catch(err => console.log(err))
-}
+import { routes } from "./urls"
+import Welcome from "./components/welcome"
+import Home from "./components/home"
 
 function App() {
   return (
     <div className="App">
-      <header className="App-header">
-        <img src={logo} className="App-logo" alt="logo" />
-        <p>
-          Edit <code>src/App.js</code> and save to reload.
-        </p>
-        <div>
-        <label htmlFor='char-input'>How many characters does </label>
-        <input id='char-input' type='text' placeholder="my string"/><span> </span>
-        <button onClick={handleSubmit}>have?</button>
-        <div>
-          <h3 id='char-count' data-testid="char-count"> </h3>
-        </div>
+      <Router>
+      <ul>
+          <li>
+            <Link to="/">Home</Link>
+          </li>
+          <li>
+            <Link to="/hello">Hello</Link>
+          </li>
+          <li>
+            <Link to="/create-quiz">Create Quiz</Link>
+          </li>
+        </ul>
+
+        <hr />
+        <Routes>
+          <Route
+            // exact
+            path={routes.root}
+            // path="create-quiz"
+            element={<Home />} 
+          />
+          <Route
+            // exact
+            path={routes.hello}
+            // path="create-quiz"
+            element={<Welcome />} 
+          />
+
+        </Routes>
+      </Router>
       </div>
-      </header>
-    </div>
   );
 }
