+++ frontend/src/components/welcome.js
@@ -1,0 +1,7 @@
+import React from 'react'
+
+function Welcome() {
+    return <h1>Hello Umar </h1>;
+}
+
+export default Welcome
