+++ frontend/src/urls.js
@@ -1,0 +1,5 @@
+export const routes = {
+    root: "/",
+    hello: "/hello",
+    createQuiz: "/create-quiz",
+}
